Add functionality to edit a task
Tasks are now editable after creation, in both memory and as displayed
in the DOM.
formController.js created as a module to serve all required form events
for the projectController.
Projects now maintain a reference to the 'active' task, which is the
task currently being viewed/editted. All forms are now initialised with
the active project as a parameter so that a reference is maintained
which allows access to the active task.