Remove dataset-id attribute for tasks/sections
Requirement to identify sections/tasks in the DOM by the using the
dataset HTML attribute has been removed. Instead, a reference to the
DOM element is maintained by the task and section objects.
Additionally, project objects now keep track of the 'active' section
which is defined by the section associated with the add task button.